node_key, link_key and policy_key moved down
the three keys have been pushed down into their list containers. If a
node is unique inside a nodes element, it is also unique inside a nffg
element (because each nffg has one and only one nodes element). The same
applies to links and policies.
The keyrefs instead must be left in the nffg element, because there must
be an ancestor relationship between the elements containing key and
keyref elements. In this way the scopes are equal to the nffg scope