moved resource for creating/updating policy
Previously for creating a policy a POST request was sent to
`/nffgs/{nffg_name}/policies`.

Now a PUT request is sent to `policies/{policy_name}`

PUT is idempotent, POST is not. Since policies should be replaced if
have the same name, PUT is appropriate, also because the path is chosen
by the client (corresponds to the name of the policy)