library fix for linux
on linux the following http status 500 appears

this is because the library jsr311-api-0.6.jar also contains a class javax.rs.ws.core.MediaType without the method fromValue. This library cannot be removed from the lib folder because it contains the package com.sun.localization that is needed for artifact compilation. The solution is not to include this library when copying to the target directory
Refers #20